requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pandas>=1.0.0
# For text analysis (optional, basic version)
nltk>=3.6.0
spacy>=3.0.0
transformers